% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgixr.R
\name{as.tags.imgixr}
\alias{as.tags.imgixr}
\title{Convert imgxr to a html tag}
\usage{
as.tags.imgixr(x, ...)
}
\arguments{
\item{x}{Object to be converted.}

\item{...}{Any additional parameters.}
}
\description{
Convert imgxr to a html tag
}
