% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{imgix-shiny}
\alias{imgix-shiny}
\alias{imgix_as_widget}
\alias{imgixOutput}
\alias{renderImgix}
\title{Shiny bindings for imgix}
\usage{
imgix_as_widget(x)

imgixOutput(outputId, width = "100\%", height = "400px")

renderImgix(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{x}{an object of class `imgix`}

\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a imgix}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using imgix within Shiny
applications and interactive Rmd documents.
}
