% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgixr.R
\name{imgix_download}
\alias{imgix_download}
\title{Download an image from the REST api}
\usage{
imgix_download(x, file)
}
\arguments{
\item{x}{an object of class `imgix`}

\item{file}{destination path for `download.file`}
}
\description{
Download an image from the REST api
}
