% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgixr.R
\name{imgixr}
\alias{imgixr}
\title{Fetch images from the imgix rest api}
\usage{
imgixr(
  img,
  fit = c("crop", "clamp", "clip", "facearea", "fill", "fillmax", "max", "scale"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{img}{an url to an image or a numeric. If a numeric is given,
it is treated as an image id for pexels.com}

\item{fit}{Resize fit mode. See \url{https://docs.imgix.com/apis/rendering/size/fit}.
the Default value was changed from `clip` to `crop`.}

\item{...}{API query parameters}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\description{
functions to load images using the imgixr rest api
}
\examples{
# images from gregordecillia.github.io/pipedream
imgixr(7417579)
imgixr(268854)
imgixr('https://images.pexels.com/photos/45848/kumamoto-japan-aso-cloud-45848.jpeg')
imgixr(4203587)
imgixr(2837863)

imgixr(4203587, txt = "hello", txt_size = 262, txt_color = "fff")
imgixr(4203587, rot = 45)
imgixr(4203587, blur = 50)
imgixr(2837863, sepia = 50)
imgixr(2837863, duotone = "191970,F08080", duotone_alpha = 60)

imgixr(2339009)
imgixr(2339009, invert = TRUE)

imgixr("https://images.unsplash.com/photo-1459478309853-2c33a60058e7")
imgixr("https://images.unsplash.com/photo-1523712999610-f77fbcfc3843")
imgixr("https://images.unsplash.com/photo-1485345488606-f2ab4a113932")

imgixr("https://assets.imgix.net/unsplash/hotairballoon.jpg")
}
