% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{imgixr_meta}
\alias{imgixr_meta}
\title{Get metadata from an imgixr picture}
\usage{
imgixr_meta(img)
}
\arguments{
\item{img}{an url to an image or a numeric. If a numeric is given,
it is treated as an image id for pexels.com}
}
\description{
Use the json format of the response to recieve metadata about an image
See \url{https://docs.imgix.com/apis/rendering/format/fm#json}.
}
\examples{
imgixr_meta(2422497)
}
