% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgixr.R
\name{knit_print.imgix}
\alias{knit_print.imgix}
\title{Show an image inside rmarkdown}
\usage{
knit_print.imgix(x, ...)
}
\arguments{
\item{x}{an object of class `imgxr`}

\item{...}{further parameters passed down to other methods}
}
\description{
Show an image inside rmarkdown
}
